backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
anthropic==0.34.0
httpx==0.27.0
supabase==2.7.0
redis==5.0.0
python-jose[cryptography]==3.3.0
pydantic==2.8.0
websockets==13.0
python-multipart==0.0.9
pytest==8.3.0
pytest-asyncio==0.24.0
